+++ Assets/Tests/SocketLayer/PeerTestAsClient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa5dd4ab2cb114b41a9bbe10bb2b1c74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
