+++ Assets/Tests/SocketLayer/PeerTestAsServer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf618d533771c1b4491816ad074adda0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
