+++ Assets/Tests/SocketLayer/PeerTestBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb7e61bf1b1f08f4a91a93201b6cae3a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
