+++ Assets/Tests/SocketLayer/PeerTestConnecting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f31d1dbe9f924224eaedbad4d5cc8d74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
